run: Write configuration along with stats logs
It's important to know what configuration produced a particular result.
Of course, logging the arguments to `run` isn't everything. There are
plenty of hard-coded hyperparameters. And we're not saving the seeds,
either. And how it is done now is not conveniently accessible now. But
it's better than nothing and it was easy to do.